Move LazyLogParameterSupplier to logging package
This was a mistake. It should never have gone in an
slf4j package.

Fixes #135
