+++ src/test/resources/logback.xml
@@ -15,7 +15,7 @@
     NOTE: A logger for this test class *must* be present and at DEBUG level for the test to work (since it
     is testing logging and specifically whether the lazy supplier executes or not.
     -->
-    <logger name="org.kiwiproject.slf4j.LazyLogParameterSupplierTest" level="DEBUG"/>
+    <logger name="org.kiwiproject.logging.LazyLogParameterSupplierTest" level="DEBUG"/>
 
     <root level="INFO">
         <appender-ref ref="STDOUT"/>
